requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
alembic>=1.14.0,<2.0.0
asyncpg>=0.30.0,<1.0.0
pydantic-settings>=2.6.0,<3.0.0
httpx>=0.28.0,<1.0.0
jinja2>=3.1.0,<4.0.0
argon2-cffi>=23.1.0,<24.0.0
redis>=5.2.0,<6.0.0
python-multipart>=0.0.9,<1.0.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
